+++ Solution_FrontEnd/FrontEnd.W/Models/CategoriaProductos.cs
@@ -14,7 +14,7 @@ namespace FrontEnd.W.Models
             Productos = new HashSet<Productos>();
         }
 
-        public decimal IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
         public string Categoria { get; set; }
 
         public virtual ICollection<Productos> Productos { get; set; }
