+++ Solution_FrontEnd/FrontEnd.W/Models/CorreoTienda.cs
@@ -9,10 +9,10 @@ namespace FrontEnd.W.Models
 {
     public partial class CorreoTienda
     {
-        public decimal IdCorreo { get; set; }
+        public int IdCorreo { get; set; }
         public string Correo { get; set; }
         public string Descripcion { get; set; }
-        public decimal IdTienda { get; set; }
+        public int IdTienda { get; set; }
 
         public virtual Tiendas IdTiendaNavigation { get; set; }
     }
