+++ Solution_FrontEnd/FrontEnd.W/Models/MarcaProductos.cs
@@ -14,7 +14,7 @@ namespace FrontEnd.W.Models
             Productos = new HashSet<Productos>();
         }
 
-        public decimal IdMarca { get; set; }
+        public int IdMarca { get; set; }
         public string MarcaProducto { get; set; }
 
         public virtual ICollection<Productos> Productos { get; set; }
