+++ Solution_FrontEnd/FrontEnd.W/Models/Productos.cs
@@ -9,14 +9,14 @@ namespace FrontEnd.W.Models
 {
     public partial class Productos
     {
-        public decimal IdProducto { get; set; }
+        public int IdProducto { get; set; }
         public string NombreProducto { get; set; }
         public string DescripcionPproducto { get; set; }
         public decimal Anno { get; set; }
         public decimal Valor { get; set; }
-        public decimal IdMarcaProducto { get; set; }
-        public decimal IdTienda { get; set; }
-        public decimal IdCategoria { get; set; }
+        public int IdMarcaProducto { get; set; }
+        public int IdTienda { get; set; }
+        public int IdCategoria { get; set; }
 
         public virtual CategoriaProductos IdCategoriaNavigation { get; set; }
         public virtual MarcaProductos IdMarcaProductoNavigation { get; set; }
